Add functional programming tools, which required adding an unbind method.
This patch got a little out of hand, apologies.

The bulk of the work is in Observable+FunctionalConvenience.swift, which adds a map, filter, reduce, and distinct methods for Observable, plus a free zip function. It also adds a passthrough debug method to help development.

To support this, I needed to modify Observable itself, adding a system of binding receipts which are used to unbind observers from each other. Observers also needed a way to retain the transient middle observers in a chain like observer.map(f).map(g), so there's now a bindings dictionary which allows strong references to be retained backwards along the chain, which can be broken when the last observer in the chain is unbound.

Finally, there's a bunch of testing of these new methods in a new Xcode project which uses cocoapods to refer back to the Sources directory.